// src/types/TodosStatusType.ts
export type TodosStatusType = {
    all: {
        label: string;
        rowColor: string
    };
    notStarted: {
        label: string;
        rowColor: string
    };
    inProgress: {
        label: string;
        rowColor: string
    };
    done: {
        label: string;
        rowColor: string
    };
}